# Make AliasTable.isampler type concrete