@ararslan 2017-09-28
It looks like this PR may have caused a conversion issue: https://github.com/JuliaStats/Distributions.jl/issues/671